pyproject.toml
[tool.poetry]
name = "chimera-nexus"
version = "1.0.0"
description = "Contextual Hybrid Intelligence for Monitoring, Evaluation & Risk Assessment"
authors = ["Analyst <analyst@chimera.local>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5.0"
pyyaml = "^6.0"
rich = "^13.7.0"

[tool.poetry.scripts]
nexus = "chimera_nexus.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"